pyproject.toml
[project]
name = "sidekick-ai"
version = "0.1.0"
description = "Asistente personal con LangGraph, Playwright y Gradio"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=0.3.9",
    "langchain-community>=0.3.20",
    "langchain-core>=0.3.0",
    "langchain-experimental>=0.3.4",
    "langgraph>=0.3.18",
    "lxml>=5.3.1",
    "playwright>=1.51.0",
    "gradio>=6.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "wikipedia>=1.4.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]